+++ simple_value_object/exceptions.py
@@ -2,13 +2,6 @@ class SimpleValueObjectException(Exception):
     pass
 
 
-class CannotBeChanged(SimpleValueObjectException):
-    def __init__(self):
-        super().__init__(
-            "You cannot change values from a Value Object, create a new one"
-        )
-
-
 class InvariantViolation(SimpleValueObjectException):
     pass